Rename file to be more consistent with its content.
