+++ lib/pulsar/context_error.rb
